+++ libs/openFrameworks/3d/ofModel.cpp
@@ -19,7 +19,7 @@ ofModel::~ofModel(){
 
 //--------------------------------------------------------------
 void ofModel::drawWireframe(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		
 		/*if(!meshes.at(i).bEnableIndices){
 			meshes.at(i).setUseIndices();
@@ -37,7 +37,7 @@ void ofModel::drawWireframe(){
 
 //--------------------------------------------------------------
 void ofModel::drawFaces(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		
 		/*if(!meshes.at(i).bEnableIndices){
 			meshes.at(i).setUseIndices();
@@ -55,7 +55,7 @@ void ofModel::drawFaces(){
 
 //--------------------------------------------------------------
 void ofModel::drawVertices(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		
 		/*if(!meshes.at(i).bEnableIndices){
 			meshes.at(i).setUseIndices();
@@ -73,49 +73,49 @@ void ofModel::drawVertices(){
 
 //--------------------------------------------------------------
 void ofModel::enableTexCoords(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).enableTexCoords();
 	}
 }
 
 //--------------------------------------------------------------
 void ofModel::enableNormals(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).enableNormals();
 	}
 }
 
 //--------------------------------------------------------------
 void ofModel::enableColors(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).enableColors();
 	}
 }
 
 //--------------------------------------------------------------
 void ofModel::disableTexCoords(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).enableTexCoords();
 	}
 }
 
 //--------------------------------------------------------------
 void ofModel::disableNormals(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).enableNormals();
 	}
 }
 
 //--------------------------------------------------------------
 void ofModel::disableColors(){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).enableColors();
 	}
 }
 
 //--------------------------------------------------------------
 void ofModel::setRenderMethod(meshRenderMethod m){
-	for (int i =0; i < meshes.size();i++){
+	for (int i =0; i < (int)meshes.size();i++){
 		meshes.at(i).setRenderMethod(m);
 	}
 }
